refactor : Pawn들의 중복된 로직을 추상 클래스로 추출
