+++ src/main/java/chess/domain/piece/BlackPawn.java
@@ -2,38 +2,13 @@ package chess.domain.piece;
 
 import chess.domain.board.Path;
 
-public class BlackPawn extends Piece {
-
-    private boolean moved = false;
-
+public class BlackPawn extends Pawn {
     public BlackPawn() {
         super(Color.BLACK);
     }
 
     @Override
-    public boolean canMove(Path path) {
-        boolean checkMove = checkMovable(path);
-        moved = true;
-        return checkMove;
-    }
-
-    private boolean checkMovable(Path path) {
-        if (!path.categoryNumOf(1)) {
-            return false;
-        }
-        if (path.isUpside()) {
-            return false;
-        }
-        if (path.containsDiagonal()) {
-            return path.isSizeOf(1) && path.isTargetHasEnemy();
-        }
-        if (path.isSizeOf(2)) {
-            return path.isAllEmpty() && isFirstMove();
-        }
-        return path.isSizeOf(1) && path.isAllEmpty();
-    }
-
-    private boolean isFirstMove() {
-        return !moved;
+    boolean isBackward(Path path) {
+        return path.isUpside();
     }
 }
